+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * Ability to display no login module if you select the option "( Do not show )" and publish no other module in the `engage-login` position
 * User Action Log entries translate the content type, useful for languages other than English
+* There's now a Clear button for the Comments filters in the backend
 
 **Bug fixes**